pyproject.toml
[project]
name = "evame-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "ruff>=0.8.2",
    "unidecode>=1.3.8",
]
